requirements.txt
langchain==0.0.352
python-dotenv==1.0.0
google-generativeai==0.3.1
chromadb==0.4.18 